+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -1092,6 +1092,12 @@ public class BoxTestGenerated extends AbstractBoxTest {
     }
 
     @Test
+    @TestMetadata("DifferentMapWrappingTypes.kt")
+    public void testDifferentMapWrappingTypes() {
+      runTest("compiler-tests/src/test/data/box/multibindings/DifferentMapWrappingTypes.kt");
+    }
+
+    @Test
     @TestMetadata("ElementsIntoSetProperty.kt")
     public void testElementsIntoSetProperty() {
       runTest("compiler-tests/src/test/data/box/multibindings/ElementsIntoSetProperty.kt");